backend/pyproject.toml
[project]
name = "ai-data-analyst-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-groq>=0.2.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.44",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]
